Directory cleaned and Structured
