--- Test_dir/t2.txt
@@ -1,1 +1,0 @@
-Data Deduplication
