+++ backend/HashAlgo-V.2.py
