--- db/t3.txt
@@ -1,1 +1,0 @@
-Data Deduplication Veritas
