--- db/t4.txt
@@ -1,193 +1,0 @@
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
-I am from VIT Pune
